scripts/scrapers/nflflag_scraper.py
"""
NFL FLAG Scraper
===============
Scrape NFL FLAG leagues from play.nflflag.com

The site uses a league finder tool that may require Selenium for dynamic content.

Usage:
    python nflflag_scraper.py --state CA
    python nflflag_scraper.py --zip 90210
"""

import requests
from bs4 import BeautifulSoup
import json
import re
import time
import logging
from pathlib import Path
from datetime import datetime
from typing import List, Dict

# Setup logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Configuration
OUTPUT_DIR = Path('../../scraped_data/raw')
OUTPUT_DIR.mkdir(parents=True, exist_ok=True)


class NFLFlagScraper:
    """Scraper for NFL FLAG leagues"""
    
    def __init__(self):
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'FlagFootballDirectory/1.0 (Educational purposes)',
            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8'
        })
        self.delay = 2
        self.base_url = "https://play.nflflag.com"
    
    def get_page(self, url: str):
        """Fetch a page with error handling"""
        try:
            logger.info(f"Fetching: {url}")
            response = self.session.get(url, timeout=10)
            response.raise_for_status()
            time.sleep(self.delay)
            return BeautifulSoup(response.content, 'html.parser')
        except Exception as e:
            logger.error(f"Error fetching {url}: {e}")
            return None
    
    def search_by_location(self, zip_code: str = None, state: str = None) -> List[Dict]:
        """
        Search for leagues by location
        
        Note: NFL FLAG's league finder may use JavaScript/API calls.
        This is a basic implementation - may need Selenium for full functionality.
        """
        leagues = []
        
        # Try the league finder page
        finder_url = f"{self.base_url}/"
        soup = self.get_page(finder_url)
        
        if not soup:
            logger.error("Could not load NFL FLAG league finder")
            return leagues
        
        # Look for API endpoints or league data in the page
        # This will need to be customized based on actual site structure
        
        logger.warning("NFL FLAG site likely requires Selenium for full scraping")
        logger.info("Consider using the Selenium-based approach for better results")
        
        return leagues
    
    def scrape_league_page(self, url: str) -> Dict:
        """Scrape a single league page"""
        soup = self.get_page(url)
        if not soup:
            return None
        
        try:
            # Extract league information
            # This needs to be customized based on actual page structure
            
            league = {
                'name': 'NFL FLAG League',
                'source': 'nflflag.com',
                'website': url,
                'formats': ['5v5'],  # NFL FLAG typically uses 5v5
                'contact_type': 'non-contact',
                'comp_levels': ['rec', 'competitive']
            }
            
            return league
            
        except Exception as e:
            logger.error(f"Error parsing league page {url}: {e}")
            return None


def main():
    """Main execution"""
    import argparse
    
    parser = argparse.ArgumentParser(description='Scrape NFL FLAG leagues')
    parser.add_argument('--zip', help='ZIP code to search')
    parser.add_argument('--state', help='State abbreviation (e.g., CA)')
    parser.add_argument('--output', default='nflflag', help='Output filename prefix')
    
    args = parser.parse_args()
    
    logger.info("="*60)
    logger.info("NFL FLAG SCRAPER")
    logger.info("="*60)
    logger.info("")
    logger.info("⚠️  Note: NFL FLAG uses dynamic content loading")
    logger.info("   This basic scraper may have limited results")
    logger.info("   For best results, use Selenium-based scraper")
    logger.info("")
    
    scraper = NFLFlagScraper()
    
    # Search for leagues
    leagues = scraper.search_by_location(
        zip_code=args.zip,
        state=args.state
    )
    
    if leagues:
        # Save results
        timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
        output_file = OUTPUT_DIR / f'{args.output}_{timestamp}_leagues.json'
        
        with open(output_file, 'w', encoding='utf-8') as f:
            json.dump(leagues, f, indent=2)
        
        logger.info(f"✅ Saved {len(leagues)} leagues to {output_file}")
    else:
        logger.warning("No leagues found")
        logger.info("")
        logger.info("💡 Recommendations:")
        logger.info("   1. Use Selenium scraper for NFL FLAG")
        logger.info("   2. Check play.nflflag.com manually")
        logger.info("   3. Look for their API endpoint")


if __name__ == '__main__':
    main()