scripts/scrapers/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
python-dotenv==1.0.0
python-slugify==8.0.1
supabase==2.3.0
